+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Implemented enhancements
 
 - GP2-3403 - Atlas opportunities listing page
+- GP2-3400: Template for generic atlas content page, which can be used for new How We Can Help page
+- GP2-3399+GP2-3401 - Nav fixups
 - NOTICKET - Fixed why invest in UK url on navigation
 - NOTICKET - Amend phrasing in Speak To Us panel
 - NOTICKET - change Investment sub-menu to feature sectors not industries